rename: studyMember 카멜케이스 정정 studymember 사용 및 studymember service 오타수정
