+++ src/main/java/com/grepp/spring/app/controller/api/StudyController.java
@@ -2,7 +2,7 @@ package com.grepp.spring.app.controller.api;
 
 import com.grepp.spring.app.model.applicant.dto.request.ApplicationRequest;
 import com.grepp.spring.app.model.applicant.dto.request.ApplicationResultRequest;
-import com.grepp.spring.app.model.studyMmeber.dto.response.CheckGoalResponse;
+import com.grepp.spring.app.model.studymember.dto.response.CheckGoalResponse;
 import com.grepp.spring.app.model.study.dto.request.NotificationUpdateRequest;
 import com.grepp.spring.app.model.study.dto.request.StudyCreationRequest;
 import com.grepp.spring.app.model.study.dto.request.StudySearchRequest;
@@ -10,19 +10,19 @@ import com.grepp.spring.app.model.study.dto.request.StudyUpdateRequest;
 import com.grepp.spring.app.model.chat.service.ChatService;
 import com.grepp.spring.app.model.member.dto.response.ApplicantsResponse;
 import com.grepp.spring.app.model.member.dto.response.StudyMemberResponse;
-import com.grepp.spring.app.model.studyMmeber.entity.Attendance;
+import com.grepp.spring.app.model.studymember.entity.Attendance;
 import com.grepp.spring.app.model.member.service.MemberService;
 import com.grepp.spring.app.model.study.code.Category;
 import com.grepp.spring.app.model.study.code.Status;
 import com.grepp.spring.app.model.study.dto.response.StudyCreationResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
-import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyAttendanceResponse;
-import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyGoalStatusResponse;
+import com.grepp.spring.app.model.studymember.dto.response.WeeklyAttendanceResponse;
+import com.grepp.spring.app.model.studymember.dto.response.WeeklyGoalStatusResponse;
 import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyNoticeResponse;
 import com.grepp.spring.app.model.applicant.service.ApplicantService;
-import com.grepp.spring.app.model.studyMmeber.serivce.StudyMemberService;
+import com.grepp.spring.app.model.studymember.service.StudyMemberService;
 import com.grepp.spring.app.model.study.service.StudyService;
 import com.grepp.spring.infra.response.CommonResponse;
 import com.grepp.spring.infra.util.AuthorizationUtil;
