+++ src/main/java/com/grepp/spring/app/model/member/service/MemberService.java
@@ -13,22 +13,22 @@ import com.grepp.spring.app.model.member.dto.response.MemberStudyListResponse;
 import com.grepp.spring.app.model.member.dto.response.MypageStudyInfoResponse;
 import com.grepp.spring.app.model.member.dto.response.RequiredMemberInfoResponse;
 import com.grepp.spring.app.model.member.dto.response.StudyInfoResponse;
-import com.grepp.spring.app.model.studyMmeber.entity.Attendance;
+import com.grepp.spring.app.model.studymember.entity.Attendance;
 import com.grepp.spring.app.model.member.entity.Member;
-import com.grepp.spring.app.model.studyMmeber.entity.StudyMember;
+import com.grepp.spring.app.model.studymember.entity.StudyMember;
 import com.grepp.spring.app.model.member.repository.MemberRepository;
-import com.grepp.spring.app.model.studyMmeber.repository.attendance.StudyAttendanceRepository;
-import com.grepp.spring.app.model.studyMmeber.repository.StudyMemberRepository;
+import com.grepp.spring.app.model.studymember.repository.attendance.StudyAttendanceRepository;
+import com.grepp.spring.app.model.studymember.repository.StudyMemberRepository;
 import com.grepp.spring.app.model.reward.entity.OwnItem;
 import com.grepp.spring.app.model.reward.entity.RewardItem;
 import com.grepp.spring.app.model.reward.repository.ItemSetRepository;
 import com.grepp.spring.app.model.reward.repository.OwnItemIdGetRepository;
 import com.grepp.spring.app.model.reward.repository.OwnItemRepository;
 import com.grepp.spring.app.model.reward.repository.RewardItemRepository;
-import com.grepp.spring.app.model.studyMmeber.entity.GoalAchievement;
+import com.grepp.spring.app.model.studymember.entity.GoalAchievement;
 import com.grepp.spring.app.model.study.entity.Study;
 import com.grepp.spring.app.model.study.entity.StudySchedule;
-import com.grepp.spring.app.model.studyMmeber.repository.goalAchievement.GoalAchievementRepository;
+import com.grepp.spring.app.model.studymember.repository.goalAchievement.GoalAchievementRepository;
 import com.grepp.spring.infra.error.exceptions.AlreadyCheckedAttendanceException;
 import com.grepp.spring.infra.error.exceptions.AlreadyExistException;
 import com.grepp.spring.infra.error.exceptions.NotFoundException;
