+++ src/main/java/com/grepp/spring/app/model/study/service/StudyService.java
@@ -28,14 +28,14 @@ import com.grepp.spring.app.model.study.entity.Study;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
 import com.grepp.spring.app.model.study.repository.StudyRepository;
 import com.grepp.spring.app.model.study.repository.goal.StudyGoalRepository;
-import com.grepp.spring.app.model.studyMmeber.code.StudyRole;
-import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyAchievementCount;
-import com.grepp.spring.app.model.studyMmeber.dto.response.WeeklyGoalStatusResponse;
-import com.grepp.spring.app.model.studyMmeber.entity.GoalAchievement;
-import com.grepp.spring.app.model.studyMmeber.entity.StudyMember;
-import com.grepp.spring.app.model.studyMmeber.repository.StudyMemberRepository;
-import com.grepp.spring.app.model.studyMmeber.repository.goalAchievement.GoalAchievementRepository;
-import com.grepp.spring.app.model.studyMmeber.serivce.StudyMemberService;
+import com.grepp.spring.app.model.studymember.code.StudyRole;
+import com.grepp.spring.app.model.studymember.dto.response.WeeklyAchievementCount;
+import com.grepp.spring.app.model.studymember.dto.response.WeeklyGoalStatusResponse;
+import com.grepp.spring.app.model.studymember.entity.GoalAchievement;
+import com.grepp.spring.app.model.studymember.entity.StudyMember;
+import com.grepp.spring.app.model.studymember.repository.StudyMemberRepository;
+import com.grepp.spring.app.model.studymember.repository.goalAchievement.GoalAchievementRepository;
+import com.grepp.spring.app.model.studymember.service.StudyMemberService;
 import com.grepp.spring.infra.error.exceptions.AlreadyExistException;
 import com.grepp.spring.infra.error.exceptions.EarlierDateException;
 import com.grepp.spring.infra.error.exceptions.HasNotRightException;
