+++ src/main/java/com/grepp/spring/app/model/studymember/code/StudyRole.java
@@ -1,0 +1,5 @@
+package com.grepp.spring.app.model.studymember.code;
+
+public enum StudyRole {
+    MEMBER, LEADER
+}
