+++ src/main/java/com/grepp/spring/app/model/studymember/dto/response/WeeklyAttendanceResponse.java
@@ -1,7 +1,7 @@
-package com.grepp.spring.app.model.studyMmeber.dto.response;
+package com.grepp.spring.app.model.studymember.dto.response;
 
 import com.grepp.spring.app.model.member.dto.response.AttendanceResponse;
-import com.grepp.spring.app.model.studyMmeber.entity.Attendance;
+import com.grepp.spring.app.model.studymember.entity.Attendance;
 import lombok.Getter;
 
 import java.util.List;
